requirements.txt
langdetect
matplotlib
numpy
pandas
scikit_learn
sentence_transformers
umap_learn
tqdm
seaborn
hdbscan
bertopic